Refactor code generation for call data handling
- Moved gen_calldata_type_name and gen_call_data_code functions to generator.py for better organization.
- Simplified gen_calldata_type_name logic and integrated it into the call data generation process.
- Enhanced gen_call_data_code to handle both direct-bind and structured variables more effectively.
- Updated calls to these functions in BoundVariable class to streamline code generation.
