+++ slangpy/core/callsignature.py
@@ -1,18 +1,12 @@
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 from typing import TYPE_CHECKING, Any, Optional
 
-from slangpy.core.native import AccessType, CallMode
-from slangpy.core.function import PipelineType
+from slangpy.core.native import CallMode
 
 import slangpy.bindings.typeregistry as tr
 from slangpy import ModifierID, TypeReflection
 from slangpy.bindings.marshall import BindContext, ReturnContext
-from slangpy.bindings.boundvariable import (
-    BoundCall,
-    BoundVariable,
-    BoundVariableException,
-)
-from slangpy.bindings.codegen import CodeGen
+from slangpy.bindings.boundvariable import BoundCall, BoundVariable
 from slangpy.builtin.value import NoneMarshall
 from slangpy.reflection.reflectiontypes import (
     SlangFunction,
@@ -38,7 +32,7 @@ class ResolveException(Exception):
         self.message = message
 
 
-from slangpy.core.generator import KernelGenException, generate_constants  # noqa: F401
+from slangpy.core.generator import KernelGenException, generate_constants, generate_code  # noqa: F401
 
 
 # This detects if a type is a vector with its length defined by a generic
@@ -233,338 +227,3 @@ def create_return_value_binding(context: BindContext, signature: BoundCall, retu
 
     node.call_dimensionality = context.call_dimensionality
     node.python = python_type
-
-
-def generate_code(
-    context: BindContext,
-    build_info: "FunctionBuildInfo",
-    signature: BoundCall,
-    cg: CodeGen,
-) -> None:
-    """
-    Generate Slang kernel code for the given function call signature.
-    """
-
-    # Check if we're using direct entry-point params (fast path)
-    use_entrypoint_args = context.use_entrypoint_args
-
-    # Generate the header
-    cg.add_import("slangpy")
-
-    call_data_len = context.call_dimensionality
-
-    # Get the call group size so we can see about using it when generating the
-    # [numthreads(...)] attribute. We use 1 as the default size if a call
-    # group shape has not been set, as we can use that to make things "linear".
-    # Note that when size is 1, we will still launch a group of 32 threads,
-    # but each thread is conceptually in its own group of size 1. This then
-    # leads to a linearly calculated call_id based on threadID.x and the call
-    # shape's size and strides.
-    call_group_size = 1
-    call_group_shape = build_info.call_group_shape
-    if call_group_shape is not None:
-        call_group_shape_vector = call_group_shape.as_list()
-
-        # Validate call_group_shape dimensionality and values before using them
-        if len(call_group_shape_vector) > context.call_dimensionality:
-            raise KernelGenException(
-                f"call_group_shape dimensionality ({len(call_group_shape_vector)}) must be <= "
-                f"call_shape dimensionality ({context.call_dimensionality}). "
-                f"call_group_shape cannot have more dimensions than call_shape."
-            )
-        elif len(call_group_shape_vector) < context.call_dimensionality:
-            # Call group shape size is less than the call shape size so we need to
-            # pad the call group shape with 1's to account for the missing dimensions.
-            # However, inserting at the front of the list will be inefficient, so
-            # log a debug message, giving users a chance to correct their calls.
-
-            missing_dims = context.call_dimensionality - len(call_group_shape_vector)
-
-            # Pad with 1's at the beginning
-            call_group_shape_vector = [1] * missing_dims + call_group_shape_vector
-
-        # Validate that all call_group_shape values are >= 1
-        for i, dim in enumerate(call_group_shape_vector):
-            if dim < 1:
-                raise KernelGenException(
-                    f"call_group_shape[{i}] = {dim} is invalid. "
-                    f"All call_group_shape elements must be >= 1."
-                )
-
-        # Calculate call group size as product of all dimensions
-        # Also grab the group strides here as that will allow us
-        # to use the group shape as constants to improve perf
-        call_group_strides = []
-        for dim in call_group_shape_vector[::-1]:
-            call_group_strides.append(call_group_size)
-            call_group_size *= dim
-        call_group_strides.reverse()
-
-        # Check if call_group_size exceeds hardware limits
-        if call_group_size > 1024:
-            raise KernelGenException(
-                f"call_group_size ({call_group_size}) exceeds the typical 1024 maximum "
-                f"enforced by most APIs. Consider reducing your call_group_shape dimensions."
-            )
-
-    cg.add_import(build_info.module.name)
-
-    # Generate constants if specified
-    generate_constants(build_info, cg)
-
-    # Generate additional link time constants definition code. These are declared in callshape.slang
-    # and used to generated call_ids that can be queried by user modules.
-    cg.constants.append_statement(f"export static const int call_data_len = {call_data_len}")
-    cg.constants.append_statement(f"export static const int call_group_size = {call_group_size}")
-
-    # Also generate the call group shape and stride arrays as link time constants. Using constants
-    # should yield better performance than passing these in as uniforms.
-    cg.constants.append_line(f"export static const int[call_data_len] call_group_strides = {{")
-    cg.constants.inc_indent()
-    if call_group_size != 1:
-        for i in range(call_data_len):
-            cg.constants.append_line(f"{call_group_strides[i]},")
-    cg.constants.dec_indent()
-    cg.constants.append_statement("}")
-
-    cg.constants.append_line(f"export static const int[call_data_len] call_group_shape_vector = {{")
-    cg.constants.inc_indent()
-    if call_group_size != 1:
-        for i in range(call_data_len):
-            cg.constants.append_line(f"{call_group_shape_vector[i]},")
-    cg.constants.dec_indent()
-    cg.constants.append_statement("}")
-
-    # Set up code gen mode for direct args vs CallData struct
-    if use_entrypoint_args:
-        cg.skip_call_data = True
-
-    # Generate call data inputs if vector call
-    if call_data_len > 0:
-        if use_entrypoint_args:
-            # Fast path: shape arrays as individual entry-point params
-            cg.entry_point_params.append(f"uniform int[{call_data_len}] _grid_stride")
-            cg.entry_point_params.append(f"uniform int[{call_data_len}] _grid_dim")
-            cg.entry_point_params.append(f"uniform int[{call_data_len}] _call_dim")
-        else:
-            # Fallback: shape arrays inside CallData struct
-            # A group can be thought of as a "window" looking at a
-            # portion of the entire call shape. Grid here refers to the
-            # N dimensional call shape being broken up into some number of N
-            # dimensional "window"s / groups.
-            cg.call_data.append_statement(f"int[{call_data_len}] _grid_stride")
-            cg.call_data.append_statement(f"int[{call_data_len}] _grid_dim")
-            # We use the call shape dimensions to detect cases when the call shape
-            # and the call group shape are not aligned. When a thread's call id
-            # falls outside the call shape, we need it to return early. This is
-            # similar to the default linear case when the call shape size is not
-            # 32 thread aligned.
-            cg.call_data.append_statement(f"int[{call_data_len}] _call_dim")
-
-    if use_entrypoint_args:
-        cg.entry_point_params.append("uniform uint3 _thread_count")
-    else:
-        cg.call_data.append_statement("uint3 _thread_count")
-
-    # Generate call data definitions for all inputs to the kernel
-    for node in signature.values():
-        node.gen_call_data_code(cg, context)
-
-    # Get sorted list of root parameters for trampoline function
-    root_params = sorted(signature.values(), key=lambda x: x.param_index)
-
-    # Generate the trampoline function
-    trampoline_fn = "_trampoline"
-    if context.call_mode != CallMode.prim:
-        cg.trampoline.append_line("[Differentiable]")
-
-    if use_entrypoint_args:
-        # Fast path: trampoline takes individual calldata-typed params.
-        # Use __in_ prefix for param names to avoid collision with local variable names.
-        # All params are no_diff — entry-point uniforms are never differentiable.
-        # Differentiation happens through local variable assignments inside the trampoline,
-        # matching the struct-based approach where CallData was implicitly non-differentiable.
-        trampoline_params = ["Context __slangpy_context__"]
-        for x in root_params:
-            if x.create_param_block:
-                continue  # param blocks handled via _param_ at module scope
-            assert x.calldata_type_name is not None
-            arg_def = f"no_diff {x.calldata_type_name} __in_{x.variable_name}"
-            trampoline_params.append(arg_def)
-        cg.trampoline.append_line(f"void {trampoline_fn}({', '.join(trampoline_params)})")
-    else:
-        # Fallback: trampoline reads from global ParameterBlock<CallData> call_data
-        cg.trampoline.append_line(f"void {trampoline_fn}(Context __slangpy_context__)")
-    cg.trampoline.begin_block()
-
-    # Declare parameters and load inputs
-    for x in root_params:
-        assert x.vector_type is not None
-        cg.trampoline.declare(x.vector_type.full_name, x.variable_name)
-    for x in root_params:
-        if use_entrypoint_args:
-            data_name = (
-                f"_param_{x.variable_name}" if x.create_param_block else f"__in_{x.variable_name}"
-            )
-        else:
-            data_name = (
-                f"_param_{x.variable_name}"
-                if x.create_param_block
-                else f"call_data.{x.variable_name}"
-            )
-        gen_load = getattr(x.python, "gen_trampoline_load", None)
-        if gen_load is not None and gen_load(cg.trampoline, x, data_name, x.variable_name):
-            continue
-        if x.access[0] == AccessType.read or x.access[0] == AccessType.readwrite:
-            cg.trampoline.append_statement(
-                f"{data_name}.__slangpy_load(__slangpy_context__.map(_m_{x.variable_name}), {x.variable_name})"
-            )
-
-    cg.trampoline.append_indent()
-    if any(x.variable_name == "_result" for x in root_params):
-        cg.trampoline.append_code(f"_result = ")
-
-    # Get function name, if it's the init function, use the result type
-    func_name = build_info.name
-    if func_name == "$init":
-        results = [x for x in root_params if x.variable_name == "_result"]
-        assert len(results) == 1
-        assert results[0].vector_type is not None
-        func_name = results[0].vector_type.full_name
-    elif len(root_params) > 0 and root_params[0].variable_name == "_this":
-        func_name = f"_this.{func_name}"
-
-    # Get the parameters that are not the result or this reference
-    normal_params = [
-        x for x in root_params if x.variable_name != "_result" and x.variable_name != "_this"
-    ]
-
-    # Internal call to the actual function
-    cg.trampoline.append_code(
-        f"{func_name}(" + ", ".join(x.variable_name for x in normal_params) + ");\n"
-    )
-
-    # For each writable trampoline parameter, potentially store it
-    for x in root_params:
-        if (
-            x.access[0] == AccessType.write
-            or x.access[0] == AccessType.readwrite
-            or x.access[1] == AccessType.read
-        ):
-            if use_entrypoint_args:
-                data_name = (
-                    f"_param_{x.variable_name}"
-                    if x.create_param_block
-                    else f"__in_{x.variable_name}"
-                )
-            else:
-                data_name = (
-                    f"_param_{x.variable_name}"
-                    if x.create_param_block
-                    else f"call_data.{x.variable_name}"
-                )
-            gen_store = getattr(x.python, "gen_trampoline_store", None)
-            if gen_store is not None and gen_store(cg.trampoline, x, data_name, x.variable_name):
-                continue
-            if not x.python.is_writable:
-                raise BoundVariableException(f"Cannot read back value for non-writable type", x)
-            cg.trampoline.append_statement(
-                f"{data_name}.__slangpy_store(__slangpy_context__.map(_m_{x.variable_name}), {x.variable_name})"
-            )
-
-    cg.trampoline.end_block()
-    cg.trampoline.append_line("")
-
-    # Generate the main function
-    if build_info.pipeline_type == PipelineType.compute:
-        cg.kernel.append_line('[shader("compute")]')
-        if call_group_size != 1:
-            cg.kernel.append_line(f"[numthreads({call_group_size}, 1, 1)]")
-        else:
-            cg.kernel.append_line("[numthreads(32, 1, 1)]")
-        # Note: While flat_call_thread_id is 3-dimensional, we consider it "flat" and 1-dimensional because of the
-        #       true call group shape of [x, 1, 1] and only use the first dimension for the call thread id.
-        if use_entrypoint_args:
-            # Fast path: build compute_main signature with individual entry-point params
-            sig_parts = ["int3 flat_call_thread_id: SV_DispatchThreadID"]
-            # Only include SV_GroupID/SV_GroupIndex when call_data_len > 0
-            # (they feed init_thread_local_call_shape_info which isn't called otherwise)
-            if call_data_len > 0:
-                sig_parts.append("int3 flat_call_group_id: SV_GroupID")
-                sig_parts.append("int flat_call_group_thread_id: SV_GroupIndex")
-            sig_parts.extend(cg.entry_point_params)
-            cg.kernel.append_line(f"void compute_main({', '.join(sig_parts)})")
-        else:
-            # Fallback: no uniform params (reads from global ParameterBlock<CallData>)
-            cg.kernel.append_line(
-                "void compute_main(int3 flat_call_thread_id: SV_DispatchThreadID, int3 flat_call_group_id: SV_GroupID, int flat_call_group_thread_id: SV_GroupIndex)"
-            )
-    elif build_info.pipeline_type == PipelineType.ray_tracing:
-        cg.kernel.append_line('[shader("raygen")]')
-        if use_entrypoint_args:
-            sig_parts = list(cg.entry_point_params)
-            cg.kernel.append_line(f"void raygen_main({', '.join(sig_parts)})")
-        else:
-            cg.kernel.append_line("void raygen_main()")
-    else:
-        raise RuntimeError(f"Unknown pipeline type: {build_info.pipeline_type}")
-
-    cg.kernel.begin_block()
-
-    if build_info.pipeline_type == PipelineType.ray_tracing:
-        cg.kernel.append_statement("int3 flat_call_thread_id = DispatchRaysIndex();")
-
-    # Bounds check — use _thread_count directly in fast path, call_data._thread_count in fallback
-    if use_entrypoint_args:
-        cg.kernel.append_statement("if (any(flat_call_thread_id >= _thread_count)) return")
-    else:
-        cg.kernel.append_statement(
-            "if (any(flat_call_thread_id >= call_data._thread_count)) return"
-        )
-
-    # Loads / initializes call id
-    context_args = "flat_call_thread_id"
-
-    # Call init_thread_local_call_shape_info to initialize the call shape info. See
-    # definition in callshape.slang.
-    if call_data_len > 0:
-        # In fast path, shape arrays are direct entry-point params; in fallback, prefixed with call_data.
-        grid_prefix = "" if use_entrypoint_args else "call_data."
-        if build_info.pipeline_type == PipelineType.compute:
-            cg.kernel.append_line(
-                f"""
-    if (!init_thread_local_call_shape_info(flat_call_group_thread_id,
-        flat_call_group_id, flat_call_thread_id, {grid_prefix}_grid_stride,
-        {grid_prefix}_grid_dim, {grid_prefix}_call_dim))
-        return;"""
-            )
-        elif build_info.pipeline_type == PipelineType.ray_tracing:
-            cg.kernel.append_line(
-                f"""
-    if (!init_thread_local_call_shape_info(0,
-        uint3(0), flat_call_thread_id, {grid_prefix}_grid_stride,
-        {grid_prefix}_grid_dim, {grid_prefix}_call_dim))
-        return;"""
-            )
-        context_args += ", CallShapeInfo::get_call_id().shape"
-
-    cg.kernel.append_statement(f"Context __slangpy_context__ = {{{context_args}}}")
-
-    # Call the trampoline function
-    fn = trampoline_fn
-    if context.call_mode == CallMode.bwds:
-        fn = f"bwd_diff({fn})"
-
-    if use_entrypoint_args:
-        # Fast path: pass individual entry-point param names to the trampoline
-        trampoline_args = ["__slangpy_context__"]
-        for x in root_params:
-            if x.create_param_block:
-                continue  # param blocks are at module scope
-            trampoline_args.append(x.variable_name)
-        cg.kernel.append_statement(f"{fn}({', '.join(trampoline_args)})")
-    else:
-        # Fallback: trampoline reads from global call_data
-        cg.kernel.append_statement(f"{fn}(__slangpy_context__)")
-
-    cg.kernel.end_block()
